add Choose button for monster special loot item (#731)
