+++ src/dialogxml/dialogs/strchoice.hpp
@@ -12,6 +12,7 @@
 #include <string>
 #include <vector>
 #include <functional>
+#include <boost/optional.hpp>
 #include "dialog.hpp"
 #include "dialogxml/widgets/ledgroup.hpp"
 
@@ -77,5 +78,10 @@ public:
 	/// Get the list of strings.
 	std::vector<std::string> getStrings() const { return strings; }
 };
+
+// Prompt the player/designer for a number, which might be an index in a given list of strings.
+// Specify cancel_value to show a cancel button, which will return the given value (for example, -1)
+// Specify extra_led and led_output to show a labeled LED which will assign led_output with its status unless the dialog is canceled
+short get_num_response(short min, short max, std::string prompt, std::vector<std::string> choice_names = {}, boost::optional<short> cancel_value = boost::none, short initial_value = 0, std::string extra_led = "", bool* led_output = nullptr);
 
 #endif
