+++ src/game/boe.items.hpp
@@ -35,10 +35,6 @@ void reset_item_max();
 short item_val(cItem item);
 void place_treasure(location where,short level,short loot,short mode);
 std::string get_text_response(std::string prompt = "", pic_num_t pic = 16);
-// Prompt the player for a number, which might be an index in a given list of strings.
-// Specify cancel_value to show a cancel button, which will return the given value (for example, -1)
-// Specify extra_led and led_output to show a labeled LED which will assign led_output with its status unless the dialog is canceled
-short get_num_response(short min, short max, std::string prompt, std::vector<std::string> choice_names = {}, boost::optional<short> cancel_value = boost::none, short initial_value = 0, std::string extra_led = "", bool* led_output = nullptr);
 
 enum class eSelectPC {
 	ANY,
